APPEALS-7004 broke up appellantnotification into separate files
